+++ flink-tools/src/main/java/io/pravega/flinktools/StreamToFileJob.java
@@ -14,7 +14,9 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.pravega.client.stream.StreamCut;
 import io.pravega.connectors.flink.FlinkPravegaReader;
+import io.pravega.flinktools.util.AscendingCounterProcessFunction;
 import io.pravega.flinktools.util.ComparableRow;
+import io.pravega.flinktools.util.ExtractKeyAndCounterFromJson;
 import org.apache.flink.api.common.functions.FilterFunction;
 import org.apache.flink.api.common.functions.FlatMapFunction;
 import org.apache.flink.api.common.functions.RichFilterFunction;
@@ -84,82 +86,11 @@ public class StreamToFileJob extends AbstractJob {
                     .name("pravega-reader");
 
             final ObjectMapper objectMapper = new ObjectMapper();
-
-//            lines.flatMap((FlatMapFunction<String, Tuple3<String, Tuple, Long>>) (s, collector) -> {
-//                final JsonNode node = objectMapper.readTree(s);
-//                final Tuple t = Tuple.newInstance(1);
-//                t.setField(node.get("sensorId").asText(), 0);
-//                collector.collect(Tuple3.of(s, t, node.get("timestamp").asLong()));
-//            })
-//            .returns(new TypeHint<Tuple3<String, Tuple, Long>>(){})
-//            .keyBy(1).process(new KeyedProcessFunction<Tuple, Tuple3<String, Tuple, Long>, String>() {
-//                @Override
-//                public void processElement(Tuple3<String, Tuple, Long> value, Context ctx, Collector<String> out) throws Exception {
-//
-//                }
-//            }).printToErr();
-
-//            lines.flatMap(new FlatMapFunction<String, Tuple3<String,Row,Long>>() {
-//                @Override
-//                public void flatMap(String s, Collector<Tuple3<String,Row,Long>> collector) throws Exception {
-//                    final JsonNode node = objectMapper.readTree(s);
-//                    final Row row = new Row(1);
-//                    row.setField(0, node.get("sensorId").asText());
-//                    collector.collect(Tuple3.of(s, row, node.get("timestamp").asLong()));
-//                }
-//            }).keyBy(1).process(new KeyedProcessFunction<Tuple, Tuple3<String, Row, Long>, String>() {
-//                @Override
-//                public void processElement(Tuple3<String, Row, Long> value, Context ctx, Collector<String> out) throws Exception {
-//
-//                }
-//            }).printToErr();
-
             final DataStream<Tuple3<String,ComparableRow,Long>> withDups = lines
-                    .flatMap((FlatMapFunction<String, Tuple3<String,ComparableRow, Long>>) (s, collector) -> {
-                final JsonNode node = objectMapper.readTree(s);
-                final ComparableRow row = new ComparableRow(2);
-                row.setField(0, node.get("sensorId").asText());
-                row.setField(1, node.get("sensorId").asText());
-                collector.collect(Tuple3.of(s, row, node.get("timestamp").asLong()));
-            })
-            .returns(new TypeHint<Tuple3<String, ComparableRow, Long>>() {});
-
-//            .keyBy(1).process(new KeyedProcessFunction<Tuple, Tuple3<String, ComparableRow, Long>, String>() {
-//                @Override
-//                public void processElement(Tuple3<String, ComparableRow, Long> value, Context ctx, Collector<String> out) throws Exception {
-//                    out.collect(value.f0);
-//                }
-//            }).printToErr();
-
-//            final DataStream<Tuple3<String,String,Long>> withDups = lines.flatMap(new FlatMapFunction<String, Tuple3<String,String,Long>>() {
-//                @Override
-//                public void flatMap(String s, Collector<Tuple3<String,String,Long>> collector) throws Exception {
-//                    final JsonNode node = objectMapper.readTree(s);
-//                    collector.collect(Tuple3.of(s, node.get("sensorId").asText(), node.get("timestamp").asLong()));
-//                }
-//            });
+                    .flatMap(new ExtractKeyAndCounterFromJson(new String[]{"sensorId","sensorId"}, "timestamp", objectMapper));
             final DataStream<String> withoutDups = withDups
                     .keyBy(1)
-                    .process(new KeyedProcessFunction<Tuple, Tuple3<String, ComparableRow, Long>, Tuple3<String, ComparableRow, Long>>() {
-                        private ValueState<Long> maxCounterState;
-
-                        @Override
-                        public void open(Configuration parameters) throws Exception {
-                            maxCounterState = getRuntimeContext().getState(new ValueStateDescriptor<Long>("maxCounter", Long.class));
-                        }
-
-                        @Override
-                        public void processElement(Tuple3<String, ComparableRow, Long> value, Context ctx, Collector<Tuple3<String, ComparableRow, Long>> out) throws Exception {
-                            final long counter = value.f2;
-                            final Long maxCounter = maxCounterState.value();
-                            if (maxCounter == null || maxCounter < counter) {
-                                maxCounterState.update(counter);
-                                out.collect(value);
-                            } else {
-                                log.info("Dropping record with decreasing counter {} and key {}", counter, ctx.getCurrentKey());
-                            }
-                        }
-                    })
+                    .process(new AscendingCounterProcessFunction())
                     .map(t -> t.f0);
             final DataStream<String> toOutput = withoutDups;
             toOutput.printToErr();
