--- flink-tools/src/main/java/io/pravega/flinktools/util/AscendingCounterFilterFunction.java
@@ -1,6 +1,0 @@
-package io.pravega.flinktools.util;
-
-import org.apache.flink.api.common.functions.RichFilterFunction;
-
-public class AscendingCounterFilterFunction extends RichFilterFunction {
-}
